Fix LEDs off by one problem in OctoWS2811Ext.cpp
The Beta 1.02 hex file as posted on GitHub works fine. However compiling the source code with the original rewrite of the OctoWS2811 library results in the LEDs in a particular strip being "off by one". The resolution was to go back to the original OctoWS2811 library and pull across the code into the OctoWS2811Ext files.